+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- TextMesh Pro dependency
+- ActionCode UI System dependency
+- LanguageSelector component
+- LocalizedStringExtension
 
 ## [1.4.0] - 2025-10-11
 ### Added